requirements.txt
google-generativeai
python-dotenv
Flask
PyYAML
